setting up routes using express router
